test-jaeger/requirements.txt
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-jaeger
jaeger-client==4.6.0